Split up and clean the Seed test
It's now Derivation and GetAddress test, with the latter also handling 
getMeritHolderNick and consistency when rebooting/loading the wallet.
