--- e2e/Tests/RPC/Personal/GetMeritHolderTest.py
@@ -1,7 +1,0 @@
-"""
-"""
-
-"""
-getMeritHolderKey
-getMeritHolderNick
-"""
